billing: Extract PlanCard to its own component
